move view files to /views
